refactor: moves tests to respective source-platform folder
